requirements.txt
crewai==0.28.8
crewai_tools==0.1.6
langchain_community==0.0.29
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
